requirements.txt
python-dotenv
langchain
langchain-community
pypdf
langchain_openai 
langchain-chroma 
langchain_postgres
chromadb
streamlit